refactor: split index logic into core/* and rename internal helpers
